[C] Reorganize content block modules
