+++ client/src/backend/components/content-block/Builder/Block/parts/Delete.js
