+++ client/src/backend/components/content-block/Builder/Block/types/Hero.js
