+++ client/src/backend/components/content-block/Builder/index.js
@@ -2,19 +2,19 @@ import React, { PureComponent } from "react";
 import PropTypes from "prop-types";
 import AvailableSection from "./sections/Available";
 import CurrentSection from "./sections/Current";
-import DraggableEventHelper from "./helpers/draggableEvent";
+import DraggableEventHelper from "../helpers/draggableEvent";
 import { contentBlocksAPI, requests } from "api";
 import { DragDropContext } from "react-beautiful-dnd";
 import withConfirmation from "hoc/with-confirmation";
 import lh from "helpers/linkHandler";
 import { entityStoreActions } from "actions";
-import configHelper from "./helpers/configurations";
+import configHelper from "../helpers/configurations";
 import cloneDeep from "lodash/cloneDeep";
 
 const { request } = entityStoreActions;
 
 export class ProjectContent extends PureComponent {
-  static displayName = "Project.Content";
+  static displayName = "ContentBlock.Builder";
 
   static propTypes = {
     project: PropTypes.object,
