+++ client/src/backend/components/content-block/Builder/sections/Available.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import Header from "./parts/Header";
 import Block from "../Block";
 import { Droppable } from "react-beautiful-dnd";
-import resolver from "../helpers/resolver";
+import resolver from "../../helpers/resolver";
 
 export default class ProjectContentSectionsAvailable extends PureComponent {
   static displayName = "Project.Content.Sections.Available";
