+++ client/src/backend/components/content-block/TypeForm/types/Hero.js
