+++ client/src/backend/components/content-block/helpers/configurations.js
