+++ client/src/backend/components/content-block/helpers/draggableEvent.js
