+++ client/src/backend/components/content-block/helpers/resolver.js
@@ -1,4 +1,4 @@
-import blocks from "../Block/types";
+import blocks from "../Builder/Block/types";
 import forms from "../TypeForm/types";
 import contentBlockHelpers from "helpers/contentBlockHelpers";
 import transform from "lodash/transform";
