+++ client/src/backend/components/content-block/index.js
@@ -1,0 +1,7 @@
+import Builder from "./Builder";
+import TypeForm from "./TypeForm";
+
+export default {
+  Builder,
+  TypeForm
+};
