+++ client/src/backend/components/project/index.js
@@ -1,9 +1,7 @@
-import Content from "./Content";
 import Form from "./form";
 import ListItem from "./ListItem";
 
 export default {
-  Content,
   Form,
   ListItem
 };
