--- client/src/backend/containers/project/content/index.js
@@ -1,5 +1,0 @@
-import Form from "./Form";
-
-export default {
-  Form
-};
