+++ client/src/frontend/components/content-block/Block/index.js
