+++ client/src/frontend/components/content-block/Block/types/Hero.js
