+++ client/src/frontend/components/content-block/Block/types/Metadata.js
