+++ client/src/frontend/components/content-block/Block/types/RecentActivity.js
