+++ client/src/frontend/components/content-block/helpers/resolver.js
