+++ client/src/frontend/components/content-block/index.js
